refactor: reorganize tiger-related modules and enhance download functionality
