+++ src/tiger/download.ts
@@ -3,8 +3,9 @@ import { log, revealLog } from "../logger";
 import * as os from "os";
 import fs from "fs";
 import path from "path";
-import { downloadFile } from "./downloadFile";
-import { exec, ExecOptions } from "child_process";
+import { downloadFile } from "../utils/downloadFile";
+import { execAsync } from "../utils/exec";
+import { ensureTigerDir } from "../config/ensureTigerPath";
 
 /**
  * Download the latest version of ck3-tiger from GitHub.
@@ -23,31 +24,42 @@ export async function downloadTiger(): Promise<string | undefined> {
 
         const downloadUrl = findPlatformAsset(latestRelease.assets, platform);
         if (!downloadUrl) {
-            vscode.window.showErrorMessage("No ck3-tiger binary found for your platform");
+            vscode.window.showErrorMessage(
+                "No ck3-tiger binary found for your platform"
+            );
             return undefined;
         }
 
-        const tigerPath = await ensureTigerPath();
+        const tigerPath = await ensureTigerDir();
         const fileExtension = path.extname(downloadUrl);
-        const fileName = fileExtension === ".zip" ? "tiger.zip" : "tiger.tar.gz";
+        const fileName =
+            fileExtension === ".zip" ? "tiger.zip" : "tiger.tar.gz";
         archivePath = path.join(tigerPath, fileName);
 
         await downloadFile(downloadUrl, archivePath).then(() => {
             log(`File downloaded as ${fileName}.`);
         });
 
-        await platformSpecificExtractArchive(archivePath, tigerPath, platform).then(() => {
+        await platformSpecificExtractArchive(
+            archivePath,
+            tigerPath,
+            platform
+        ).then(() => {
             log("File extracted successfully.");
         });
 
-        const executableFile = platform === "win32" ? "ck3-tiger.exe" : "ck3-tiger";
+        const executableFile =
+            platform === "win32" ? "ck3-tiger.exe" : "ck3-tiger";
         const executablePath = path.join(tigerPath, executableFile);
 
         return executablePath;
     } catch (error: unknown) {
-        const errorMessage = error instanceof Error ? error.message : String(error);
+        const errorMessage =
+            error instanceof Error ? error.message : String(error);
         log(`Error downloading tiger: ${errorMessage}`);
-        vscode.window.showErrorMessage(`Failed to download ck3-tiger: ${errorMessage}`);
+        vscode.window.showErrorMessage(
+            `Failed to download ck3-tiger: ${errorMessage}`
+        );
         return undefined;
     } finally {
         if (archivePath && fs.existsSync(archivePath)) {
@@ -66,18 +78,24 @@ async function fetchLatestRelease(): Promise<any | undefined> {
 
     try {
         const octokit = new Octokit();
-        const response = await octokit.request("GET /repos/{owner}/{repo}/releases/latest", {
-            owner: "amtep",
-            repo: "tiger", // amtep renamed "ck3-tiger" to "tiger" :(
-            headers: {
-                "X-GitHub-Api-Version": "2022-11-28",
-            },
-        });
+        const response = await octokit.request(
+            "GET /repos/{owner}/{repo}/releases/latest",
+            {
+                owner: "amtep",
+                repo: "tiger", // amtep renamed "ck3-tiger" to "tiger" :(
+                headers: {
+                    "X-GitHub-Api-Version": "2022-11-28",
+                },
+            }
+        );
         return response.data;
     } catch (error: unknown) {
-        const errorMessage = error instanceof Error ? error.message : String(error);
+        const errorMessage =
+            error instanceof Error ? error.message : String(error);
         log(`Error fetching latest release: ${errorMessage}`);
-        vscode.window.showErrorMessage(`Failed to fetch latest ck3-tiger release: ${errorMessage}`);
+        vscode.window.showErrorMessage(
+            `Failed to fetch latest ck3-tiger release: ${errorMessage}`
+        );
         return undefined;
     }
 }
@@ -88,12 +106,24 @@ async function fetchLatestRelease(): Promise<any | undefined> {
  * @param {NodeJS.Platform} platform - The current platform.
  * @returns {Object | undefined} Object containing download URL or undefined if not found.
  */
-export function findPlatformAsset(assets: any[], platform: NodeJS.Platform): string | undefined {
+export function findPlatformAsset(
+    assets: any[],
+    platform: NodeJS.Platform
+): string | undefined {
     log("Latest release assets:");
-    log(JSON.stringify(assets.filter(asset => asset.name.includes("ck3")), null, 4));
+    log(
+        JSON.stringify(
+            assets.filter((asset) => asset.name.includes("ck3")),
+            null,
+            4
+        )
+    );
 
     const platformName = platform === "win32" ? "windows" : "linux";
-    const currentAsset = assets.find((asset) => asset.name.includes("ck3") && asset.name.includes(platformName));
+    const currentAsset = assets.find(
+        (asset) =>
+            asset.name.includes("ck3") && asset.name.includes(platformName)
+    );
 
     if (!currentAsset) {
         return undefined;
@@ -102,50 +132,22 @@ export function findPlatformAsset(assets: any[], platform: NodeJS.Platform): str
     return currentAsset.browser_download_url;
 }
 
-async function ensureTigerPath(): Promise<string> {
-    const context = await vscode.commands.executeCommand("getContext") as vscode.ExtensionContext;
-
-    // Ensure the global storage directory exists
-    const globalStoragePath = context.globalStorageUri.fsPath;
-    if (!fs.existsSync(globalStoragePath)) {
-        fs.mkdirSync(globalStoragePath, { recursive: true });
-    }
-
-    const tigerPath = path.join(globalStoragePath, "ck3-tiger");
-    if (!fs.existsSync(tigerPath)) {
-        fs.mkdirSync(tigerPath, { recursive: true });
-        log(`Created directory: ${tigerPath}`);
-    }
-
-    return tigerPath;
-}
-
 async function platformSpecificExtractArchive(
     archivePath: string,
     destDir: string,
-    platform: NodeJS.Platform,
+    platform: NodeJS.Platform
 ): Promise<{
     stdout: string;
     stderr: string;
 }> {
     switch (platform) {
         case "win32":
-            return execAsync(`tar -xf "${archivePath}" -C "${destDir}"`, { shell: "powershell.exe" });
+            return execAsync(`tar -xf "${archivePath}" -C "${destDir}"`, {
+                shell: "powershell.exe",
+            });
         case "linux":
             return execAsync(`tar -xzf "${archivePath}" -C "${destDir}"`);
         default:
             throw new Error("Unsupported platform for extraction");
     }
-}
-
-export function execAsync(command: string, options?: ExecOptions): Promise<{ stdout: string; stderr: string }> {
-    return new Promise((resolve, reject) => {
-        exec(command, options, (error, stdout, stderr) => {
-            if (error) {
-                reject(error);
-            } else {
-                resolve({ stdout: stdout.toString(), stderr: stderr.toString() });
-            }
-        });
-    });
 }
