+++ src/ui/statusBar.ts
